Enable a version of xfail
This requires using a custom Errorf and Fatalf
throughout, instead of returning errors, and
using Logf instead of the core Errorf and Fatalf
if Xfail is true on the Case.

This is because the testing package does not
provide a way of doing expected fail.

The result is that an expected fail will pass
when it fails and fail when it passes, but we
don't really know that it is an expected fail.