Add Accepts to response handlers
This allows us to exit early (with errorf not fatalf)
if the response body does not match the the accept
criteria (which is effectively the content-type of
the response).

We will want to allow response handlers to register
themselves, but that comes later.

This finishses self.yaml, a new file will be added
for the next round.